+++ server/src/models/skill.rs
@@ -1,3 +1,4 @@
+use bigdecimal::BigDecimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -12,10 +13,10 @@ pub struct SkillData {
     pub level_required: i32,
     pub mp_cost: i32,
     pub cooldown: i32,
-    pub cast_time: f32,
-    pub range: f32,
+    pub cast_time: BigDecimal,
+    pub range: BigDecimal,
     #[serde(rename = "area_of_effect")]
-    pub area_of_effect: f32,
+    pub area_of_effect: BigDecimal,
     pub effects: Vec<serde_json::Value>,
     pub learnable_by: Vec<String>,
 }
